+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `network.giantswarm.io/wildcard-cname-target` annotation to override the wildcard CNAME target for workload clusters.
+
 ## [0.25.0] - 2024-06-03
 
 ### Added